requirements.txt
Django==1.11.24
django-crispy-forms==1.7.2
django-model-utils==3.1.2 #Django model mixins and utilities.
django-sortedm2m==1.5.0
django-anymail==0.3.1
django-docs==0.3.0
djangorestframework==3.9.4
matplotlib==2.0.0
numpy==1.19.1
python-dateutil==2.8.0
pytz==2019.1
redis==3.3.4
requests==2.22.0
django-grappelli==2.11.1
pandas==1.1.0
scipy==1.1.0
django-smart-selects==1.5.4 #for adding filter or group (chained) models in dropdowns
django-extensions==2.1.9
django-cors-headers==2.4.0
django-taggit==0.23.0
django-ses==1.0.0 #previously used version 0.8.12
Pillow==3.1.0
Sphinx==1.5.2 #for creating documentation
IPython==5.3.0
jupyter==1.0.0
django-redis==4.10.0
xlrd==1.1.0 #for reading data and formatting information from Excel files
hiredis==1.0.0
django-admin-rangefilter==0.3.8 #for selecting date range in django admin site
django-autocomplete-light==3.2.10 #for giving autocomplete feature for a large number of records
graphene-django==2.3.2
django-filter==1.1.0 #used by graphene-django for filtering on fields, python2.7 last version
gunicorn[gthread]==19.9.0
tqdm==4.32.2
jupyterlab==0.33.12
psycopg2-binary==2.8.5
xlsxwriter==1.1.8
nbformat==4.4
boto3==1.11.10
